Create FavoriteItem comp
